LayoutTests/media/modern-media-controls/status-label/status-label-macos.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing <code>StatusText</code> properties when used in <code>MacOSInlineMediaControls</code>.");

const mediaControls = new MacOSInlineMediaControls({ width: 600, height: 300 });
mediaControls.statusLabel.enabled = true;
mediaControls.statusLabel.text = "Hello World";

document.body.appendChild(mediaControls.element);

let style;
shouldBecomeEqual("document.body.contains(mediaControls.statusLabel.element)", "true", () => {
    style = getComputedStyle(mediaControls.statusLabel.element);

    shouldBeEqualToString("style.top", "14.5px");
    shouldBeEqualToString("style.mixBlendMode", "plus-lighter");
    shouldBeEqualToRGBAColor("style.color", "rgba(255, 255, 255, 0.45)");

    mediaControls.element.remove();
    debug("");
    finishJSTest();
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
